Move funcs from morpheusapi to morpheus_funcs
* Moved functions from the morpheusapi module to the morpheus_funcs module
* Import the entire morpheus_funcs module into dependent modules
* Remove some unused imports
