+++ plugins/module_utils/morpheus_funcs.py
@@ -12,9 +12,9 @@ author: James Riach
 
 import re
 try:
-    from morpheusapi import MorpheusApi, dict_filter
+    from morpheusapi import MorpheusApi
 except ModuleNotFoundError:
-    from ansible_collections.morpheus.core.plugins.module_utils.morpheusapi import MorpheusApi, dict_filter
+    from ansible_collections.morpheus.core.plugins.module_utils.morpheusapi import MorpheusApi
 
 
 def class_to_dict(input_class: object) -> dict:
@@ -148,6 +148,112 @@ def dict_compare_equality(dict_a: dict, dict_b: dict) -> bool:
             return False
 
     return True
+
+
+def dict_filter(dictionary: dict, filter_keys: list) -> dict:
+    """Return a filtered dictionary based on a list of keys to retain
+
+    Args:
+        dictionary (dict): The original dictionary to filter
+        filter_keys (list): A list of keys to retain.
+          To filter subkeys use a list or tuple in the format of ['key', ['subkeyA', 'subkeyB']]
+
+    Returns:
+        dict: A dictionary with only the keys specified in filter_keys
+    """
+    subkey_filters = {k: v for k, v in enumerate(filter_keys) if isinstance(v, (list, set, tuple))}
+
+    if len(subkey_filters) > 0:
+        for k, v in subkey_filters.items():
+            filter_keys[k] = v[0]
+
+    filtered_dict = {k: v for k, v in dictionary.items() if k in filter_keys}
+
+    if len(subkey_filters) > 0:
+        for _, v in subkey_filters.items():
+            try:
+                if isinstance(filtered_dict[v[0]], (list, set, tuple)):
+                    for idx, _ in enumerate(filtered_dict[v[0]]):
+                        filtered_dict[v[0]][idx] = dict_filter(filtered_dict[v[0]][idx], v[1])
+                    continue
+                filtered_dict[v[0]] = dict_filter(filtered_dict[v[0]], v[1])
+            except KeyError:
+                continue
+
+    return filtered_dict
+
+
+def dict_keys_to_camel_case(dictionary: dict, recursive: bool = True) -> dict:
+    """Convert keys of a dictionary to camelCase format
+
+    Args:
+        dictionary (dict): The dictionary to convert
+        recursive (bool, optional): Recursively convert subkeys. Defaults to True.
+
+    Returns:
+        dict: Dictionary with keys in camelCase
+    """
+    camel_dict = {}
+
+    for k in dictionary.keys():
+        camel_case_key = ''.join(c.capitalize() for c in k.lower().split('_'))
+        lower_camel_case_key = camel_case_key[0].lower() + camel_case_key[1:]
+        camel_dict[lower_camel_case_key] = dictionary[k]
+
+        if recursive and isinstance(camel_dict[lower_camel_case_key], dict):
+            if isinstance(camel_dict[lower_camel_case_key], list):
+                camel_dict[lower_camel_case_key] = [dict_keys_to_camel_case(item) if isinstance(item, dict) else item
+                                                    for item in camel_dict[lower_camel_case_key]]
+                continue
+            if isinstance(camel_dict[lower_camel_case_key], set):
+                camel_dict[lower_camel_case_key] = {dict_keys_to_camel_case(item) if isinstance(item, dict) else item
+                                                    for item in camel_dict[lower_camel_case_key]}
+                continue
+            if isinstance(camel_dict[lower_camel_case_key], tuple):
+                camel_dict[lower_camel_case_key] = (dict_keys_to_camel_case(item) if isinstance(item, dict) else item
+                                                    for item in camel_dict[lower_camel_case_key])
+                continue
+            if isinstance(camel_dict[lower_camel_case_key], dict):
+                camel_dict[lower_camel_case_key] = dict_keys_to_camel_case(camel_dict[lower_camel_case_key])
+                continue
+
+    return camel_dict
+
+
+def dict_keys_to_snake_case(dictionary: dict, recursive: bool = True) -> dict:
+    """Convert keys of a dictionary to snake_case format
+
+    Args:
+        dictionary (dict): The dictionary to convert
+        recursive (bool, optional): Recursively convert subkeys. Defaults to True.
+
+    Returns:
+        dict: Dictionary with keys in snake_case format
+    """
+    snake_dict = {}
+
+    for k in dictionary.keys():
+        new_key_name = re.sub('((?<=[a-z0-9])[A-Z]|(?!^)(?<!_)[A-Z](?=[a-z]))', r'_\1', k).lower()
+        snake_dict[new_key_name] = dictionary[k]
+
+        if recursive:
+            if isinstance(snake_dict[new_key_name], list):
+                snake_dict[new_key_name] = [dict_keys_to_snake_case(item) if isinstance(item, dict) else item
+                                            for item in snake_dict[new_key_name]]
+                continue
+            if isinstance(snake_dict[new_key_name], set):
+                snake_dict[new_key_name] = {dict_keys_to_snake_case(item) if isinstance(item, dict) else item
+                                            for item in snake_dict[new_key_name]}
+                continue
+            if isinstance(snake_dict[new_key_name], tuple):
+                snake_dict[new_key_name] = (dict_keys_to_snake_case(item) if isinstance(item, dict) else item
+                                            for item in snake_dict[new_key_name])
+                continue
+            if isinstance(snake_dict[new_key_name], dict):
+                snake_dict[new_key_name] = dict_keys_to_snake_case(snake_dict[new_key_name])
+                continue
+
+    return snake_dict
 
 
 def instance_filter(morpheus_api: MorpheusApi, module_params: dict, key_filter: list = None) -> list:
